+++ release-notes.md
@@ -1,4 +1,5 @@
 Changes since v2.0.1 include:
+- Airspy: use http transport for stats - allowing stage2 to get access to the stats as well
 - stage2: fix route api settings
 - stage2: more tuning of the stage2 home page
 - micro feeder: don't show empty string as listener